rename utils -> util directory. and add triple slash ts reference directive for declaration files
