+++ packages/zk-eddsa-ticket-pcd/src/index.ts
@@ -1,1 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
+/// <reference path="./util/declarations/circomlibjs.d.ts" />
+/// <reference path="./util/declarations/snarkjs.d.ts" />
 export * from "./ZKEdDSATicketPCD";
