+++ packages/zk-eddsa-ticket-pcd/src/util/utils.ts
