+++ packages/zk-eddsa-ticket-pcd/test/EdDSATicketPrivatePCD.spec.ts
@@ -5,6 +5,9 @@ import {
   SemaphoreIdentityPCDPackage,
   SemaphoreIdentityPCDTypeName
 } from "@pcd/semaphore-identity-pcd";
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
+/// <reference path="../src/util/declarations/circomlibjs.d.ts" />
+/// <reference path="../src/util/declarations/snarkjs.d.ts" />
 import { Identity } from "@semaphore-protocol/identity";
 import { expect } from "chai";
 import "mocha";
